Added Goal Progress

+++ features/goals/src/main/java/app/chamich/feature/goals/model/Goal.kt
@@ -10,8 +10,8 @@ data class Goal(
     override val id: Long = 0L,
     override val title: String,
     override val measuredIn: Int,
-    override val totalEffort: Long,
-    override val progress: Long,
+    override val totalEffort: Int,
+    override val progress: Int,
     override val completeData: Long,
     override val category: Int,
     override val color: Int
